+++ web/js/ui/cloud_auth/AccountControlDropdown.tsx
@@ -28,10 +28,10 @@ export class AccountControlDropdown extends React.PureComponent<IProps, IState> 
                         id="account-control-button"
                         size="sm"
                         onClick={() => NULL_FUNCTION}
-
+                        style={{whiteSpace: 'nowrap'}}
                         className="header-filter-clickable p-1 pl-2 pr-2 border">
 
-                    <i className="fas fa-cloud-upload-alt" style={{marginRight: '5px'}}></i>
+                    <i className="fas fa-cloud-upload-alt" style={{marginRight: '5px'}}/>
 
                     {/*<span className="d-none-mobile">*/}
                     {/*    {AppRuntime.isBrowser() ? 'Account' : 'Cloud Sync'}*/}